fix: create account name null (#5)
This fixes an issue where the canary acc service was not using the Account Name field to input data into the SQL table 'accounts'. This resulted in the creation of accounts without an Account Name, which is a key field and needs to be unique. As a result, this caused an error with the message "SQLSTATE[23000]: Integrity constraint violation: 1062 Duplicate entry '' for key 'name'".

To fix this issue, the pull request adds a validation to ensure that the Account Name field is not null before creating a new account in the database. If the Account Name is null, an error is thrown and the account creation is aborted. This ensures that only valid accounts with unique Account Names are created in the database.

Overall, this pull request improves the reliability and accuracy of the canary acc service by fixing the "Account Name" null issue.