fix: solution for checkbox (#3)
This changes addresses issue #2, which caused a problem with checkboxes that resulted in the error "ERROR: SQLSTATE[HY000]: General error: 1366 Incorrect integer value: '' for column 'account' at row 1".

The cause of the issue was due to checkboxes not being flagged when not selected, and therefore not being included in the request. While some users had success with adding a hidden input checkbox with a default value of 0, this did not work for everyone. As a result, this pull request includes a logic fix in the Controller to validate if the postVars are being set. If not, a 0 is added to the request.

While this may not be the best solution, it is currently the most effective fix for the issue. The team may continue to work on improving this solution in the future.

Additionally, this also addresses the missing status feedback on submit and is currently being worked on. Overall, this pull request improves the reliability and functionality of the service by fixing the checkbox issue.